requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
google-generativeai==0.8.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0